Create xmas_tree_multi_random_function_VF2.py
Python 3 application to control the LEDs on the Pi Hut Xmas tree for the Raspberry-pi.

My first attempt at Python so it is probably not very elegant but it seams to work.

Program for the Pi Hut Xmas Tree  that does multiple patterns in a random order
Glyn Morgan 
Version: Final-2 02/12/17

Tested in Raspbian Stretch
Needs the GPIO Zero Library

The PI Hut Xams tree board is optionally modified with a button between the otherwise unused GPIO3 & ground.
This is to provide a clean shutdown method when used in a stand-alone mode.
The button is wired between pins 5 & 6 of the GPIO connector on the tree.
It can take a while for the program to spot that the shutdown button has been pushed. So hold the button in until the light sequence stops.